+++ litho-processor/src/main/java/com/facebook/litho/specmodels/model/CachedValueValidation.java
@@ -80,13 +80,14 @@ public class CachedValueValidation {
         onCalculateCachedValueMethods) {
       for (MethodParamModel param : onCalculateCachedValueMethod.methodParams) {
         if (!(MethodParamModelUtils.isComponentContextParam(param))
+            && !(param instanceof TreePropModel)
             && !(param instanceof PropModel)
             && !(param instanceof StateParamModel)
             && !(param instanceof InjectPropModel)) {
           validationErrors.add(
               new SpecModelValidationError(
                   param.getRepresentedObject(),
-                  "@OnCalculateCachedValue methods may only take ComponentContext, Props, @InjectProps and State as params."));
+                  "@OnCalculateCachedValue methods may only take ComponentContext, @Prop, @TreeProp, @InjectProp and @State as params."));
         }
       }
     }